Make export states a default action
Added hash to grammar objects to determine whether a grammar has been
modified based on the names and order of production bodies.

This allows a states object to be created and exported as a JSON
compliant JavaScript object that can then be re-imported on subsequent
compilations. If the grammar has been found to have not changed based on
the hash, and only the production functions have been modified, then the
compilation of states can be skipped in, with the imported JSON states
used instead.

This states object, which also includes a minified grammar object, is
stored alongside the source grammar file, and is automatically loaded
when a grammar is compiled. This file is named

``<source grammar file name>.debug_data.js``

In addition, this data can be loaded into a lrParser run to assist
in debugging failed parses. If this data is present, when a parse fails
a history of states will be logged with detailed information of state
actions and token information.

-----

Modified error recovery sequence

Fix error with token not being assigned correct during a fork action

Make generated production names deterministic