Break: Use Token as the default value for parser stack instead of string slices
Rename `pos` objects to `tok` in grammar reduce functions and source files

#changelog

Replaced string slices with Token as the default value for reduced
productions.

Renamed `pos` (POSition) to `tok` (TOKen) in reduce functions.

Modified Token~slice to allow subsections of the token's slice to be
extracted, similar to the String~slice method.

Added Token~token_slice that returns a new Token instead of a sub-string.
The interface remains the same as Token~slice otherwise.