Add NULL to STRUCT_CLASSIFICATION~struct types
The null type is implied when constructing AST structs that do not
directly assign struct properties values, but instead apply a
struct type/class specifier to the struct property:
example:

```javascript
 ... { { t_Struct, child: t_ChildStruct } }
```

In the above case, the `child` property is assigned a type `t_ChildStruct`
but not a value. When a `t_Struct` is constructed by this parse action,
the `child` property should be assigned a placeholder value, thus NULL.

Adding `{ type: NULL, body: [] }` to STRUCT_CLASSIFICATION is analogous
to this result.